+++ MortgageCalculator/Models/PatchInterestRateInput.cs
@@ -4,6 +4,6 @@ namespace MortgageCalculator.Models
 {
     public class PatchInterestRateInput
     {
-        public decimal NewRate;
+        public double NewRate;
     }
 }
